+++ linearSearch.cpp
@@ -80,7 +80,7 @@ std::variant<TDicionario*, TArvBin> LinearSearch::testInsere(std::vector<int> da
 
     auto dicionario = TDicionario_Inicia(dataset.size(), counter_comparisons);
 
-    int sum = 0;
+    long long sum = 0;
 
     for (int i = 0; i < dataset.size(); ++i) {
 
@@ -104,7 +104,7 @@ std::variant<TDicionario*, TArvBin> LinearSearch::testPesquisa(std::variant<TDic
     if (std::holds_alternative<TDicionario*>(dicionario)) {
         TDicionario* dic_ptr = std::get<TDicionario*>(dicionario);
 
-        int sum = 0;
+        long long sum = 0;
 
         for (int i = 0; i < dataset.size(); ++i) {
 
@@ -128,7 +128,7 @@ std::variant<TDicionario*, TArvBin> LinearSearch::testRetira(std::variant<TDicio
 
         TDicionario* dic_ptr = std::get<TDicionario*>(dicionario);
 
-        int sum = 0;
+        long long sum = 0;
 
         for (int i = 0; i < dataset.size(); ++i) {
